service_embed_documents/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=requirements.txt --pip-args=None pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
authlib==1.6.5
    # via weaviate-client
boto3==1.40.40
    # via embed_documents (pyproject.toml)
boto3-stubs[s3]==1.40.40
    # via embed_documents (pyproject.toml)
botocore==1.40.46
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.40.33
    # via boto3-stubs
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
confluent-kafka==2.11.1
    # via embed_documents (pyproject.toml)
cryptography==46.0.2
    # via authlib
deprecation==2.1.0
    # via weaviate-client
distro==1.9.0
    # via openai
googleapis-common-protos==1.70.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.75.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   weaviate-client
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   weaviate-client
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jiter==0.11.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mypy-boto3-s3==1.40.26
    # via boto3-stubs
openai==2.2.0
    # via embed_documents (pyproject.toml)
opentelemetry-api==1.37.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.37.0
    # via embed_documents (pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.37.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.37.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.37.0
    # via opentelemetry-exporter-otlp
opentelemetry-proto==1.37.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.37.0
    # via
    #   embed_documents (pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.58b0
    # via opentelemetry-sdk
packaging==25.0
    # via deprecation
protobuf==6.32.1
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
    #   weaviate-client
pycparser==2.23
    # via cffi
pydantic==2.12.0
    # via
    #   openai
    #   weaviate-client
pydantic-core==2.41.1
    # via pydantic
pypdf2==3.0.1
    # via embed_documents (pyproject.toml)
python-dateutil==2.9.0.post0
    # via botocore
requests==2.32.5
    # via opentelemetry-exporter-otlp-proto-http
s3transfer==0.14.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
types-awscrt==0.27.6
    # via botocore-stubs
types-s3transfer==0.13.1
    # via boto3-stubs
typing-extensions==4.15.0
    # via
    #   anyio
    #   grpcio
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via
    #   botocore
    #   requests
validators==0.35.0
    # via weaviate-client
weaviate-client==4.17.0
    # via embed_documents (pyproject.toml)
zipp==3.23.0
    # via importlib-metadata
